Add tests and CI for both proxies.
There are two sets of tests: CI tests, which must run w/o TLS and DNS
for the proxies, so are basic http calls. These tests are run as landing
tests.

The second set of tests full-tests.sh, which will only run against
production, as it needs full TLS/DNS set up. But it tests the actual
docker and git clients against the proxy.