Proxy Honeycomb markers
This is not ready yet.

We need to a) authenticate the client, like we do for otel-gateway b)
force using our writekey, so an attacker couldn't write data to
a different HC account.

Other proxies all include the org in the url, so its simpler to
restrict. Honeycomb users the auth token as both auth and identity, so
it is not as simple.