Include the host name in the access logs
Because all vhosts need to log to a single stream (stdout), then we need
to include the vhost domain in the logs. This would have allowed us to
easily filter out all but docker-proxy logs last week.

Additionally, have rejigged the logformat again, in bikeshed fashion, so
the entire logged host/path?qs is copy-pasteable. Slight wrinkle that we
all ways add a ? even if there's no qs, but that's just like the Good
Ol' Days, so slight nostalgia.